# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Produtos Favoritos",
        "description": "Documentação da API RESTful para o desafio técnico do aiqfome.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API local"
        }
    ],
    "paths": {
        "/api/client/register": {
            "post": {
                "tags": [
                    "Auth - Client"
                ],
                "summary": "Registra um novo cliente",
                "operationId": "8351fa64e73547c72ea44d3cc94b256b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente registrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "client": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/client/login": {
            "post": {
                "tags": [
                    "Auth - Client"
                ],
                "summary": "Autentica um cliente",
                "operationId": "c0deeda73de83424e4ff6911539d8645",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login bem-sucedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "expires_in": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/api/client/me": {
            "get": {
                "tags": [
                    "Auth - Client"
                ],
                "summary": "Retorna o cliente autenticado",
                "operationId": "d4fd015cd38fa44c5a33ed7bd40a7bde",
                "responses": {
                    "200": {
                        "description": "Dados do cliente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/client/logout": {
            "post": {
                "tags": [
                    "Auth - Client"
                ],
                "summary": "Realiza logout do cliente",
                "operationId": "31d0941ee304ecfc80621ceaf0badc27",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/login": {
            "post": {
                "tags": [
                    "Auth - User"
                ],
                "summary": "Login do usuário do sistema (admin)",
                "operationId": "b9a6a8b44a0c416a582473741334da81",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login bem-sucedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/api/user/logout": {
            "post": {
                "tags": [
                    "Auth - User"
                ],
                "summary": "Logout do usuário do sistema",
                "operationId": "1a0569d743575459b9a86160adfb0c5f",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/me": {
            "get": {
                "tags": [
                    "Auth - User"
                ],
                "summary": "Retorna o usuário autenticado (admin)",
                "operationId": "025e25a08b9a35616e8c4b54123cb620",
                "responses": {
                    "200": {
                        "description": "Dados do usuário autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/client/favorites/{clientId}": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Lista os produtos favoritos de um cliente",
                "operationId": "16cead9200f4f28cbde1dfcda0db19a9",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de favoritos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/client/favorites/plus": {
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Incrementa (ou cria) um produto favorito",
                "operationId": "1eb68702082e3ca2a126b599130bd27e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto favoritado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Produto inválido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/client/favorites/minus": {
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Decrementa ou remove um favorito",
                "operationId": "a0a223f2e8a121d69c1d5bd3d3e767d7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Quantidade atualizada ou item removido"
                    },
                    "404": {
                        "description": "Favorito não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/client/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Lista todos os produtos da FakeStore",
                "operationId": "0378717e8e1da3ae3e602fc68549ca46",
                "responses": {
                    "200": {
                        "description": "Lista de produtos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "title": {
                                                "type": "string"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "category": {
                                                "type": "string"
                                            },
                                            "image": {
                                                "type": "string",
                                                "format": "uri"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/client/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Busca um produto por ID",
                "operationId": "79ae799ec8de014fe7e3f0051e632a29",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do produto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "category": {
                                            "type": "string"
                                        },
                                        "image": {
                                            "type": "string",
                                            "format": "uri"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Atualiza um produto via FakeStore API",
                "operationId": "5c7c04e984a801d63ea62fc17a49640b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "uri"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado com sucesso"
                    },
                    "400": {
                        "description": "Erro ao atualizar produto"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Remove um produto via FakeStore API",
                "operationId": "de4684b1798f0a46e809cd8f51b26759",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto removido com sucesso"
                    },
                    "400": {
                        "description": "Falha ao remover produto"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/test-swagger": {
            "get": {
                "tags": [
                    "Test"
                ],
                "summary": "Teste de documentação funcionando",
                "operationId": "4c92f594a6409c28dfdd6f08243325f1",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth - Client",
            "description": "Auth - Client"
        },
        {
            "name": "Auth - User",
            "description": "Auth - User"
        },
        {
            "name": "Favorites",
            "description": "Favorites"
        },
        {
            "name": "Products",
            "description": "Products"
        },
        {
            "name": "Test",
            "description": "Test"
        }
    ]
}